backend/pyproject.toml
[project]
name = "webhook-service"
version = "0.1.0"
description = ""
authors = [
    {name = "solalohana",email = "solalohana@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi (>=0.115.12,<0.116.0)",
    "uvicorn[standard] (>=0.34.2,<0.35.0)",
    "pydantic[email] (>=2.11.4,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "pydantic-settings (>=2.0)",
    "alembic (>=1.15.2,<2.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "sqlalchemy-utils (>=0.41.2,<0.42.0)",
    "passlib[bcrypt] (==1.7.4)",
    "bcrypt (==3.2.2)",
    "httpx (==0.27.0)",
    "boto3 (>=1.38.18,<2.0.0)",
    "moto (>=5.1.4,<6.0.0)",
    "pytest (>=8.3.5,<9.0.0)",
    "python-jose (>=3.4.0,<4.0.0)",
    "python-multipart (>=0.0.20,<0.0.21)",
    "stripe (>=12.1.0,<13.0.0)"
]

[tool.poetry]
packages = [{include = "app"}]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.5"

